use of LfAsymBus class
