+++ src/main/java/com/powsybl/openloadflow/ac/equations/asym/AbstractShuntFortescueEquationTerm.java
@@ -14,7 +14,6 @@ import com.powsybl.openloadflow.equations.AbstractElementEquationTerm;
 import com.powsybl.openloadflow.equations.Variable;
 import com.powsybl.openloadflow.equations.VariableSet;
 import com.powsybl.openloadflow.network.LfBus;
-import com.powsybl.openloadflow.network.extensions.AsymBus;
 import com.powsybl.openloadflow.util.Fortescue;
 
 import java.util.Objects;
@@ -64,19 +63,11 @@ public abstract class AbstractShuntFortescueEquationTerm extends AbstractElement
     }
 
     protected double b() {
-        AsymBus asymBus = (AsymBus) element.getProperty(AsymBus.PROPERTY_ASYMMETRICAL);
-        if (sequenceType == Fortescue.SequenceType.ZERO) {
-            return asymBus.getbZeroEquivalent();
-        }
-        return asymBus.getbNegativeEquivalent();
+        return sequenceType == Fortescue.SequenceType.ZERO ? element.getAsym().getBzEquiv() : element.getAsym().getBnEquiv();
     }
 
     protected double g() {
-        AsymBus asymBus = (AsymBus) element.getProperty(AsymBus.PROPERTY_ASYMMETRICAL);
-        if (sequenceType == Fortescue.SequenceType.ZERO) {
-            return asymBus.getgZeroEquivalent();
-        }
-        return asymBus.getgNegativeEquivalent();
+        return sequenceType == Fortescue.SequenceType.ZERO ? element.getAsym().getGzEquiv() : element.getAsym().getGnEquiv();
     }
 
 }
