+++ src/main/java/com/powsybl/openloadflow/ac/equations/asym/LoadAbcImpedantEquationTerm.java
@@ -3,8 +3,8 @@ package com.powsybl.openloadflow.ac.equations.asym;
 import com.powsybl.openloadflow.ac.equations.AcVariableType;
 import com.powsybl.openloadflow.equations.Variable;
 import com.powsybl.openloadflow.equations.VariableSet;
+import com.powsybl.openloadflow.network.LfAsymBus;
 import com.powsybl.openloadflow.network.LfBus;
-import com.powsybl.openloadflow.network.extensions.AsymBus;
 import com.powsybl.openloadflow.network.extensions.AsymBusVariableType;
 import com.powsybl.openloadflow.network.extensions.LegConnectionType;
 import com.powsybl.openloadflow.util.ComplexMatrix;
@@ -48,7 +48,7 @@ public class LoadAbcImpedantEquationTerm extends AbstractAsymmetricalLoad {
         // [Ic*]    [Sca0/Vca0² * Vca - Sbc0/Vbc0² * Vbc]
         //
 
-        AsymBus asymBus = (AsymBus) bus.getProperty(AsymBus.PROPERTY_ASYMMETRICAL);
+        LfAsymBus asymBus = bus.getAsym();
         if (asymBus == null) {
             throw new IllegalStateException("unexpected null pointer for an asymmetric bus " + bus.getId());
         }
@@ -70,9 +70,9 @@ public class LoadAbcImpedantEquationTerm extends AbstractAsymmetricalLoad {
         Complex vNegativeComplex = new Complex(negativeSequence.getX(), negativeSequence.getY());
         Complex vZeroComplex = new Complex(zeroSequence.getX(), zeroSequence.getY());
 
-        Complex vA0 = AsymBus.getVa0();
-        Complex vB0 = AsymBus.getVb0();
-        Complex vC0 = AsymBus.getVc0();
+        Complex vA0 = LfAsymBus.getVa0();
+        Complex vB0 = LfAsymBus.getVb0();
+        Complex vC0 = LfAsymBus.getVc0();
 
         boolean hasPhaseA = asymBus.isHasPhaseA();
         boolean hasPhaseB = asymBus.isHasPhaseB();
@@ -192,7 +192,7 @@ public class LoadAbcImpedantEquationTerm extends AbstractAsymmetricalLoad {
         // [dIc*]    [Sca0/Vca0² * dVca - Sbc0/Vbc0² * dVbc]
         //
 
-        AsymBus asymBus = (AsymBus) bus.getProperty(AsymBus.PROPERTY_ASYMMETRICAL);
+        LfAsymBus asymBus = bus.getAsym();
         if (asymBus == null) {
             throw new IllegalStateException("unexpected null pointer for an asymmetric bus " + bus.getId());
         }
@@ -213,9 +213,9 @@ public class LoadAbcImpedantEquationTerm extends AbstractAsymmetricalLoad {
         Complex sB = new Complex(asymBus.getPb(), asymBus.getQb());
         Complex sC = new Complex(asymBus.getPc(), asymBus.getQc());
 
-        Complex vA0 = AsymBus.getVa0();
-        Complex vB0 = AsymBus.getVb0();
-        Complex vC0 = AsymBus.getVc0();
+        Complex vA0 = LfAsymBus.getVa0();
+        Complex vB0 = LfAsymBus.getVb0();
+        Complex vC0 = LfAsymBus.getVc0();
 
         boolean hasPhaseA = asymBus.isHasPhaseA();
         boolean hasPhaseB = asymBus.isHasPhaseB();
