+++ src/main/java/com/powsybl/openloadflow/network/util/UniformValueVoltageInitializer.java
@@ -7,9 +7,9 @@
  */
 package com.powsybl.openloadflow.network.util;
 
+import com.powsybl.openloadflow.network.LfAsymBus;
 import com.powsybl.openloadflow.network.LfBus;
 import com.powsybl.openloadflow.network.LfNetwork;
-import com.powsybl.openloadflow.network.extensions.AsymBus;
 import com.powsybl.openloadflow.util.Fortescue;
 
 /**
@@ -33,7 +33,7 @@ public class UniformValueVoltageInitializer implements VoltageInitializer {
     }
 
     public static double getMagnitude(LfBus bus, Fortescue.SequenceType sequenceType) {
-        AsymBus asymBus = AsymBus.getAsymBus(bus);
+        LfAsymBus asymBus = bus.getAsym();
         if (asymBus == null) {
             return 1;
         }
@@ -41,7 +41,7 @@ public class UniformValueVoltageInitializer implements VoltageInitializer {
     }
 
     public static double getAngle(LfBus bus, Fortescue.SequenceType sequenceType) {
-        AsymBus asymBus = AsymBus.getAsymBus(bus);
+        LfAsymBus asymBus = bus.getAsym();
         if (asymBus == null) {
             return 0;
         }
